renamed checklist index-new to index
