+++ html/checklist/index.php
@@ -41,7 +41,12 @@ ob_end_flush();
 /* HEADER */ require('layout/header2.php');
 /* HEADER */ require('layout/header1.php');
 
-
+print'
+<script language="javascript" type="text/javascript" src="table.js"></script>
+<script language="javascript" type="text/javascript" src="jquery-1.12.4.js"></script>
+<script language="javascript" type="text/javascript" src="jquery.dataTables.min.js"></script>
+<link rel="stylesheet" type="text/css" href="jquery.dataTables.min.css">
+';
 
 
 print'
@@ -141,7 +146,7 @@ FROM userCardChecklist
 				SELECT
 userCardChecklist.quantity, userCardChecklist.wishlistQuantity, userCardChecklist.dateadded,
 cardList.series, cardList.cardnum, cardList.player, cardList.description, cardList.team,
-series_info.series_name, series_info.series_tag, series_info.sort
+series_info.series_name, series_info.series_tag, series_info.sort, userCardChecklist.marketSale,userCardChecklist.marketWishlist
 
 FROM userCardChecklist
 					LEFT JOIN cardList
@@ -150,12 +155,75 @@ FROM userCardChecklist
 					LEFT JOIN series_info
 						ON userCardChecklist.series = series_info.series_tag
 
-					WHERE userCardChecklist.wishlistQuantity > 0 AND userCardChecklist.userid ='".$user->id."'
+					WHERE userCardChecklist.wishlistQuantity  > 0 AND userCardChecklist.userid ='".$user->id."'
 					ORDER BY series_info.sort, userCardChecklist.cardnum
 					"
 				);
 				$wishlistQuantity = $R_cards2->num_rows;
 				$R_cards2->free();
+
+								// get count of MARKET SALES cards
+								$R_cards2 = $db_main->query("
+								SELECT
+				userCardChecklist.quantity, userCardChecklist.wishlistQuantity, userCardChecklist.dateadded,
+				cardList.series, cardList.cardnum, cardList.player, cardList.description, cardList.team,
+				series_info.series_name, series_info.series_tag, series_info.sort, userCardChecklist.marketSale,userCardChecklist.marketWishlist
+				
+				FROM userCardChecklist
+									LEFT JOIN cardList
+										ON cardList.series = userCardChecklist.series
+										AND cardList.cardnum = userCardChecklist.cardnum
+									LEFT JOIN series_info
+										ON userCardChecklist.series = series_info.series_tag
+				
+									WHERE  userCardChecklist.marketSale > 0 AND userCardChecklist.userid ='".$user->id."'
+									ORDER BY series_info.sort, userCardChecklist.cardnum
+									"
+								);
+								$MarketSaleQuantity = $R_cards2->num_rows;
+								$R_cards2->free();
+
+
+												// get count of MARKET WISHLIST cards
+												$R_cards2 = $db_main->query("
+												SELECT
+								userCardChecklist.quantity, userCardChecklist.wishlistQuantity, userCardChecklist.dateadded,
+								cardList.series, cardList.cardnum, cardList.player, cardList.description, cardList.team,
+								series_info.series_name, series_info.series_tag, series_info.sort, userCardChecklist.marketSale,userCardChecklist.marketWishlist
+								
+								FROM userCardChecklist
+													LEFT JOIN cardList
+														ON cardList.series = userCardChecklist.series
+														AND cardList.cardnum = userCardChecklist.cardnum
+													LEFT JOIN series_info
+														ON userCardChecklist.series = series_info.series_tag
+								
+													WHERE userCardChecklist.marketWishlist > 0 AND userCardChecklist.userid ='".$user->id."'
+													ORDER BY series_info.sort, userCardChecklist.cardnum
+													"
+												);
+												$MarketWishlistQuantity = $R_cards2->num_rows;
+												$R_cards2->free();
+
+
+
+
+				$now = time();
+				$lastBump = $now - 14*24*60*60;
+				// 14 days in the past, to compare to last bump -- can bump every 2 weeks
+
+
+				// get count of MARKET SALES LASTBUMP cards
+				$R_cards2 = $db_main->query("SELECT lastBumpDate FROM marketSale WHERE userid='".$user->id."' AND expired = 'N' AND $lastBump-lastBumpDate  > 0");
+				$MarketSalesBump = $R_cards2->num_rows;
+				$R_cards2->free();
+
+
+				// get count of MARKET WISHLIST LAST BUMP cards
+				$R_cards2 = $db_main->query("SELECT lastBumpDate FROM marketWishlist WHERE userid='".$user->id."' AND expired = 'N' AND $lastBump-lastBumpDate  > 0");
+				$MarketWishlistBump = $R_cards2->num_rows;
+				$R_cards2->free();
+
 
 
 
@@ -175,7 +243,7 @@ FROM userCardChecklist
 						 and add them to your personal checklist! It looks like your ebay username
 						 is "'.$ebayID.'". Unfortunately, we don\'t see any auctions
 						 tied to your username. If your information is correct, please send us an email and we
-						 can look into this. To update your upsername,
+						 can look into this. To update your username,
 						 <a href="https://helmarbrewing.com/account/">click to add your ebay username on the account
 						  info page.</a> Please note, in the future if you win any of our card auctions,
 						  come back and we can import those auctions to your personal checklist!
@@ -203,7 +271,7 @@ FROM userCardChecklist
 			}
 
 
-			if($checklistQuantity+$wishlistQuantity==0){
+			if($checklistQuantity+$wishlistQuantity+$MarketSaleQuantity+$MarketWishlistQuantity==0){
 				print '<p>You have no cards in your Checklist and Wishlist!
 				<a href="https://helmarbrewing.com/artwork/">You can add cards by heading to the Artwork pages</a>.
 				</p>';
@@ -212,7 +280,7 @@ FROM userCardChecklist
 									$R_cards = $db_main->query("
 
 									SELECT
-					userCardChecklist.quantity, userCardChecklist.wishlistQuantity, userCardChecklist.dateadded,
+					userCardChecklist.quantity, userCardChecklist.wishlistQuantity, userCardChecklist.marketSale,userCardChecklist.marketWishlist,userCardChecklist.card_note, userCardChecklist.dateadded,
 					cardList.series, cardList.cardnum, cardList.player, cardList.description, cardList.team,
 					series_info.series_name, series_info.series_tag, series_info.sort
 
@@ -223,28 +291,53 @@ FROM userCardChecklist
 										LEFT JOIN series_info
 											ON userCardChecklist.series = series_info.series_tag
 
-										WHERE userCardChecklist.quantity + userCardChecklist.wishlistQuantity > 0 AND userCardChecklist.userid ='".$user->id."'
+										WHERE userCardChecklist.quantity + userCardChecklist.wishlistQuantity + userCardChecklist.marketSale + userCardChecklist.marketWishlist > 0 AND userCardChecklist.userid ='".$user->id."'
 										ORDER BY series_info.sort, userCardChecklist.cardnum
 										"
 									);
+									
+									print '<p>';
+									if($MarketSaleQuantity>0){
+										if($MarketSalesBump>0){
+											print '
+											<a href="javascript:;" title="Click to bump your Marketplace Sales items to the top of the Marketplace page and extend your expiration date"><img src="/img/mktsalesactive.png" alt="Helmar Brewing" onclick="renewMarketSale()"></a>
+											';
+										}else{
+											print '
+											<img src="/img/mktsalesinactive.png" alt="Helmar Brewing" title="You can only renew every 14 days">
+											';
+										}
+									}
+									if($MarketWishlistQuantity>0){
+										if($MarketWishlistBump>0){
+											print '
+											<a href="javascript:;" title="Click to bump your Marketplace Wanted items to the top of the Marketplace page and extend your expiration date"><img src="/img/mktwishlistactive.png" alt="Helmar Brewing" onclick="renewMarketWish()"></a>
+											';
+										}else{
+											print '
+											<img src="/img/mktwishlistinactive.png" alt="Helmar Brewing" title="You can only renew every 14 days">
+											';
+										}
+									}
+									print '</p>';
 
 					            print'
-					                    <table class="tables">
-					                      <thead>
-					                        <tr>
+					                    <table id="chklist" class="display compact" width="100%" cellspacing="0">
+										<thead>
+										<tr>
+										<th>Card Series</th>
+										<th>Card #</th>
+										<th>Player</th>
+										<th>Stance / Position</th>
+										<th>Team</th>
+										<th>Stock Pictures</th>
+										<th title="Use the checkbox to indicate you own this card">Card Owned <i class="fa fa-info-circle"></i></th>
+										<th title="Use the checkbox to indicate you have an interest in this card when it shows up on eBay. Checking the box, you will receive an email when it goes live">eBay Auction Wishlist <i class="fa fa-info-circle"></i></th>
+										<th title="Use the checkbox to indicate you own this card and wish to sell/trade on the Marketplace Use the comment icon to note any notable items suchs as card condition, quality, and artwork (some Helmar cards have different artwork)">Sell Card on Marketplace <i class="fa fa-info-circle"></i></th>
+										<th title="Use the checkbox to indicate you want to buy/trade for this card on the Marketplace">Want from Marketplace <i class="fa fa-info-circle"></i></th>
+										</tr>
+									  </thead>
 
-																		<th>Card Series</th>
-					                          <th>Card Number</th>
-					                          <th>Player</th>
-					                          <th>Stance / Position</th>
-					                          <th>Team</th>
-
-					                          <th>Pictures</th>
-																		<th>Checklist</th>
-																		<th>Wishlist</th>
-																		<th>Active Auction</th>
-					                        </tr>
-					                      </thead>
 					                      <tbody>
 					            ';
 
@@ -271,29 +364,18 @@ FROM userCardChecklist
 					                    print'
 					                        <tr>
 
-																			<td>'.$card->series_name.'</td>
+												<td>'.$card->series_name.'</td>
 					                            <td>'.$card->cardnum.'</td>
 					                            <td>'.$card->player.'</td>
 					                            <td>'.$card->description.'</td>
 					                            <td>'.$card->team.'</td>
 
 					                    ';
-					                    if($card->averagesold == 0){
-					                  //      print'
-					                  //          <td></td>
-					                  //          <td></td>
-					                  //      ';
-					                    }else{
-
-					                //        print'
-					                  //          <td>'.$card->lastsold.'</td>
-					                  //          <td>'.$card->maxSold.'</td>
-					                  //      ';
-					                    }
+					                   
 
 					                    // need to add ************ for one row, check if picture exists
 					                    print'
-					                            <td align="center" class="picCol">
+					                            <td>
 					                    ';
 
 
@@ -351,8 +433,31 @@ FROM userCardChecklist
 											print '<td><a href="javascript:;" ><i class="fa fa-square-o" onclick="wishlist(\''.$card->series.'\',\''.$card->cardnum.'\')" id="WISH'.$card->cardnum.'_'.$card->series.'"></i></a></td>';
 										}
 
-											// add ebay auction if active
-												print '<td><a href="http://www.ebay.com/itm/'. $wishlist[$card->series_tag.'-'.$card->cardnum]['auctionID'].'/" target="_blank">'.$wishlist[$card->series_tag.'-'.$card->cardnum]['auctionID'].'</a></td>';
+										if($card->quantity > 0){	
+											if($card->marketSale > 0){
+												print '<td><a href="javascript:;" ><i class="fa fa-check-square-o" onclick="marketSale(\''.$card->series.'\',\''.$card->cardnum.'\')" id="mktSale'.$card->cardnum.'_'.$card->series.'"></i></a> &nbsp;';
+												if($card->card_note <> ""){
+													print '<a href="javascript:;" title="'.$card->card_note.'"><i class="fa fa-comments" onclick="marketSaleComment(\''.$card->series.'\',\''.$card->cardnum.'\',\''.$card->card_note.'\')" id="mktSaleComment'.$card->cardnum.'_'.$card->series.'"></i></a>';
+												}else{
+													print '<a href="javascript:;" title="'.$card->card_note.'"><i class="fa fa-comment-o" onclick="marketSaleComment(\''.$card->series.'\',\''.$card->cardnum.'\',\''.$card->card_note.'\')" id="mktSaleComment'.$card->cardnum.'_'.$card->series.'"></i></a>';
+												}
+												print '</td>';
+											} else{
+												print '<td><a href="javascript:;" ><i class="fa fa-square-o" onclick="marketSale(\''.$card->series.'\',\''.$card->cardnum.'\')" id="mktSale'.$card->cardnum.'_'.$card->series.'"></i></a></td>';
+											}
+										} else{
+											print '<td></td>';
+										}
+
+
+										if($card->marketWishlist > 0){
+											print '<td><a href="javascript:;" ><i class="fa fa-check-square-o" onclick="marketWishlist(\''.$card->series.'\',\''.$card->cardnum.'\')" id="mktWish'.$card->cardnum.'_'.$card->series.'"></i></a></td>';
+										} else{
+											print '<td><a href="javascript:;" ><i class="fa fa-square-o" onclick="marketWishlist(\''.$card->series.'\',\''.$card->cardnum.'\')" id="mktWish'.$card->cardnum.'_'.$card->series.'"></i></a></td>';
+										}
+
+
+
 
 					            		/* end row */
 					            		print '</tr>';
@@ -364,7 +469,7 @@ FROM userCardChecklist
 													$R_cards2->free();
 					            }else{
 					                print'
-					                    <tr><td colspan="8">could not get list of cards</td></tr>
+					                    <tr><td colspan="10">could not get list of cards</td></tr>
 					                ';
 					            }
 					            print'
@@ -409,6 +514,52 @@ print'
 $db_auth->close();
 $db_main->close();
 ?>
+
+
+<script>
+$(document).ready(function() {
+    $('#chklist').DataTable({
+  "columns": [
+    { "width": "13%" },
+    { "width": "5%" },
+    { "width": "18%" },
+    { "width": "10%" },
+    { "width": "10%" },
+	{ "width": "10%" }, 
+	{ "width": "7%" },
+    { "width": "9%" },
+    { "width": "9%" },
+	{ "width": "9%" }
+  ],
+	//"scrollY": "600px",
+	//"scrollCollapse": true,
+	//"scrollX": false,
+	"pageLength": 50,
+	initComplete: function () {
+					 this.api().columns().every( function () {
+							 var column = this;
+							 var select = $('<select><option value=""></option></select>')
+									 .appendTo( $(column.footer()).empty() )
+									 .on( 'change', function () {
+											 var val = $.fn.dataTable.util.escapeRegex(
+													 $(this).val()
+											 );
+
+											 column
+													 .search( val ? '^'+val+'$' : '', true, false )
+													 .draw();
+									 } );
+
+							 column.data().unique().sort().each( function ( d, j ) {
+									 select.append( '<option value="'+d+'">'+d+'</option>' )
+							 } );
+					 } );
+			 }
+},
+);
+} );
+
+</script>
 
 <script>
 function checklist(s, c){
@@ -477,6 +628,97 @@ function wishlist(s, c){
 </script>
 
 <script>
+function marketSale(s, c){
+
+	var txt = "To list a card on the Helmar Brewing Marketplace, you must agree to the following:\n\nYou understand that you are listing a card or cards on the Helmar Brewing Marketplace.\nWithin the Marketplace, your identity will be anonymous. In using the Marketplace, \nanother user may or may not reach out to you via the Marketplace contact form. You \nwill receive this communication to the email you have listed under your \nhelmarbrewing.com account. You are free to communicate with the other party \nas you wish. This will take place outside of the helmarbrewing.com website and \nyou will not hold Helmar Brewing responsible for any issues that may occur outside \nof the helmarbrewing.com website.\n\nHelmar Brewing recommends safe trading practices.\n\nIn order to list your card, you must agree to the above terms. If you do not agree, \nyou will be taken back to the page you were last visiting.\n\nClick \'ok\' to agree to these terms, or click \'cancel\' to return to the checklist";
+
+    document.getElementById('fullscreenload').style.display = 'block';
+    $.get(
+        "/artwork/ajax/marketSale/",
+		{ series:s, cardnum:c },
+        function( data ) {
+			if(data.error === 0){
+				if(data.qty === '1'){
+					if(window.confirm(txt)){
+						$('#mktSale' + c + '_' + s).removeClass('fa-square-o');
+						$('#mktSale' + c + '_' + s).addClass('fa-check-square-o');
+						location.reload();
+					}else{
+						$.get(
+							"/artwork/ajax/marketSale/",
+							{ series:s, cardnum:c },
+							"json"
+						);
+						window.alert("You need to accept terms to list on the Marketplace");
+					}
+				}else if(data.qty === '0'){
+						$('#mktSale' + c + '_' + s).removeClass('fa-check-square-o');
+						$('#mktSale' + c + '_' + s).addClass('fa-square-o');
+						location.reload();
+				} else {
+					alert("else part...");
+				}
+			}else{
+				alert(data.msg);
+				location.reload();
+			}
+            document.getElementById('fullscreenload').style.display = 'none';
+        },
+        "json"
+    )
+    .fail(function() {
+        alert('There was an error, refresh the page.');
+    });
+	
+	
+
+}
+</script>
+
+<script>
+function marketWishlist(s, c){
+	var txt = "To list a card on the Helmar Brewing Marketplace, you must agree to the following:\n\nYou understand that you are listing a card or cards on the Helmar Brewing Marketplace.\nWithin the Marketplace, your identity will be anonymous. In using the Marketplace, \nanother user may or may not reach out to you via the Marketplace contact form. You \nwill receive this communication to the email you have listed under your \nhelmarbrewing.com account. You are free to communicate with the other party \nas you wish. This will take place outside of the helmarbrewing.com website and \nyou will not hold Helmar Brewing responsible for any issues that may occur outside \nof the helmarbrewing.com website.\n\nHelmar Brewing recommends safe trading practices.\n\nIn order to list your card, you must agree to the above terms. If you do not agree, \nyou will be taken back to the page you were last visiting.\n\nClick \'ok\' to agree to these terms, or click \'cancel\' to return to the checklist";
+    document.getElementById('fullscreenload').style.display = 'block';
+    $.get(
+        "/artwork/ajax/marketWishlist/",
+		{ series:s, cardnum:c },
+        function( data ) {
+			if(data.error === 0){
+				if(data.qty === '1'){
+					if(window.confirm(txt)){
+						$('#mktWish' + c + '_' + s).removeClass('fa-square-o');
+						$('#mktWish' + c + '_' + s).addClass('fa-check-square-o');
+						location.reload();
+					}else{
+						$.get(
+							"/artwork/ajax/marketWishlist/",
+							{ series:s, cardnum:c },
+							"json"
+						);
+						window.alert("You need to accept terms to list on the Marketplace");
+					}
+				}else if(data.qty === '0'){
+						$('#mktWish' + c + '_' + s).removeClass('fa-check-square-o');
+						$('#mktWish' + c + '_' + s).addClass('fa-square-o');
+						location.reload();
+				} else {
+					alert("else part...");
+				}
+			}else{
+				alert(data.msg);
+				location.reload();
+			}
+            document.getElementById('fullscreenload').style.display = 'none';
+        },
+        "json"
+    )
+    .fail(function() {
+        alert('There was an error, refresh the page.');
+    });
+}
+</script>
+
+<script>
 function ebayImport(s){
     document.getElementById('fullscreenload').style.display = 'block';
     $.get(
@@ -511,7 +753,94 @@ function ebayImport(s){
 
 
 }
+</script>
 
+<script>
+function marketSaleComment(s, c, t){
 
+	var t = prompt("Please enter any comments about your card that you would share with others", t);
+
+    document.getElementById('fullscreenload').style.display = 'block';
+   $.get(
+        "/artwork/ajax/marketSaleComment/",
+		{ series:s, cardnum:c, comment:t },
+//        function( data ) {
+//			if(data.error === 0){
+//				if(data.qty === '1'){
+//					if(window.confirm(txt)){
+//					//	$('#mktSaleComment' + c + '_' + s).removeClass('fa-square-o');
+//					//	$('#mktSaleComment' + c + '_' + s).addClass('fa-check-square-o');
+//						location.reload();
+//					}else{
+//						$.get(
+//							"/artwork/ajax/marketSale/",
+//							{ series:s, cardnum:c },
+//							"json"
+//						);
+//						window.alert("You need to accept terms to list on the Marketplace");
+//					}
+//				}else if(data.qty === '0'){
+//					//	$('#mktSaleComment' + c + '_' + s).removeClass('fa-check-square-o');
+//					//	$('#mktSaleComment' + c + '_' + s).addClass('fa-square-o');
+//						location.reload();
+//				} else {
+//					alert("else part...");
+//				}
+//			}else{
+//				alert(data.msg);
+//				location.reload();
+//			}
+//            document.getElementById('fullscreenload').style.display = 'none';
+//        },
+        "json"
+    )
+    .fail(function() {
+        alert('There was an error, refresh the page.');
+    });
+
+	document.getElementById('fullscreenload').style.display = 'none';
+	alert("Card comment updated");
+	location.reload();
+
+}
+</script>
+
+<script>
+function renewMarketSale(){
+
+document.getElementById('fullscreenload').style.display = 'block';
+$.get(
+	"/checklist/ajax/renewMarketSale/",
+	"json"
+)
+.fail(function() {
+	alert('There was an error, refresh the page.');
+});
+
+document.getElementById('fullscreenload').style.display = 'none';
+alert("All current Market sales cards have been renewed");
+location.reload();
+
+}
+
+</script>
+
+<script>
+function renewMarketWish(){
+
+document.getElementById('fullscreenload').style.display = 'block';
+$.get(
+	"/checklist/ajax/renewMarketWant/",
+	"json"
+)
+.fail(function() {
+	alert('There was an error, refresh the page.');
+});
+
+document.getElementById('fullscreenload').style.display = 'none';
+alert("All current Market wanted cards have been renewed");
+location.reload();
+
+}
 
 </script>
